--- server/src/dal/usersHandler.ts
@@ -1,8 +1,0 @@
-import { IUser, User } from "../models/UserModel";
-
-const findOne = async (data: Partial<IUser>): Promise<IUser | null> => {
-  return User.findOne({
-    data,
-  });
-};
-export default { findOne };
