+++ server/tsconfig.json
@@ -26,7 +26,7 @@
   "exclude": ["node_modules/*"],
   "include": [
     "./src/**/*.ts",
-    //"./tests/**/*.ts",
+    // "./tests/**/*.ts", // TODO: This line must be included but it is causing failures while building due to types mismatch.
     "./src/db/util/*.ts"
   ]
 }
